Added error logging and improved validation. Increased the max length of a few fields.

+++ BookReviews.Impl/Logic/ReviewLogic.cs
@@ -5,8 +5,10 @@ using BookReviews.Impl.Repositories.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
+using Exception = System.Exception;
 
 namespace BookReviews.Impl.Logic
 {
@@ -14,13 +16,16 @@ namespace BookReviews.Impl.Logic
     {
         private IMapper _mapper;
         private IBookLogic _bookLogic;
+        private IExceptionLogic _exceptionLogic;
         private IGenreRepository _genreRepository;
         private IReviewRepository _reviewRepository;
 
-        public ReviewLogic(IMapper mapper, IBookLogic bookLogic, IGenreRepository genreRepository, IReviewRepository reviewRepository)
+        public ReviewLogic(IMapper mapper, IBookLogic bookLogic, IExceptionLogic exceptionLogic, 
+            IGenreRepository genreRepository, IReviewRepository reviewRepository)
         {
             _mapper = mapper;
             _bookLogic = bookLogic;
+            _exceptionLogic = exceptionLogic;
             _genreRepository = genreRepository;
             _reviewRepository = reviewRepository;
         }
@@ -29,12 +34,19 @@ namespace BookReviews.Impl.Logic
         {
             try
             {
+                if (reviewId == 0)
+                    throw new Exception("Review ID is invalid");
+
                 Entities.Review reviewRecord = await _reviewRepository.GetReview(reviewId);
                 Models.Review reviewResult = _mapper.Map<Models.Review>(reviewRecord);
 
                 if (includeBook)
                 {
                     Models.Book bookRecord = await _bookLogic.GetBook(reviewRecord.Book.Id);
+
+                    if (bookRecord == null)
+                        throw new Exception("Failed to retrieve book");
+
                     reviewResult.Book = bookRecord;
                 }
 
@@ -42,7 +54,12 @@ namespace BookReviews.Impl.Logic
             }
             catch (Exception ex)
             {
-                // log error here
+                var arguments = new Dictionary<string, string>();
+                arguments.Add("ReviewId", reviewId.ToString());
+                arguments.Add("IncludeBook", includeBook.ToString());
+
+                await _exceptionLogic.LogException(ex, "Get review error", arguments);
+                
                 return null;
             }
         }
@@ -51,6 +68,9 @@ namespace BookReviews.Impl.Logic
         {
             try
             {
+                if (bookId == 0 || bookId.ToString().Length < Globals.MaxLengths.Book.Id)
+                    throw new Exception("Book ID is invalid");
+
                 List<Entities.Review> reviewRecords = await _reviewRepository.GetReviews(bookId);
                 List<Models.Review> reviewResults = _mapper.Map<List<Models.Review>>(reviewRecords);
 
@@ -58,7 +78,11 @@ namespace BookReviews.Impl.Logic
             }
             catch (Exception ex)
             {
-                // log error here
+                var arguments = new Dictionary<string, string>();
+                arguments.Add("BookId", bookId.ToString());
+
+                await _exceptionLogic.LogException(ex, "Get reviews for book error", arguments);
+
                 return new List<Models.Review>();
             }
         }
@@ -67,6 +91,9 @@ namespace BookReviews.Impl.Logic
         {
             try
             {
+                if (userId == 0)
+                    throw new Exception("User ID is invalid");
+
                 DateTime? startDate = null;
                 DateTime? endDate = null;
 
@@ -99,7 +126,13 @@ namespace BookReviews.Impl.Logic
             }
             catch (Exception ex)
             {
-                // log error here
+                var arguments = new Dictionary<string, string>();
+                arguments.Add("UserId", userId.ToString());
+                arguments.Add("IncludeBooks", includeBooks.ToString());
+                arguments.Add("Year", year.ToString());
+
+                await _exceptionLogic.LogException(ex, "Get reviews for user error", arguments);
+
                 return new List<Models.Review>();
             }
         }
@@ -108,6 +141,9 @@ namespace BookReviews.Impl.Logic
         {
             try
             {
+                if (userId == 0)
+                    throw new Exception("User ID is invalid");
+
                 List<DateTime> reviewDateRecords = await _reviewRepository.GetReviewDates(userId);
                 List<int> reviewYears = reviewDateRecords.Select(x => x.Year).Distinct().ToList();
 
@@ -115,7 +151,11 @@ namespace BookReviews.Impl.Logic
             }
             catch (Exception ex)
             {
-                // log error here
+                var arguments = new Dictionary<string, string>();
+                arguments.Add("UserId", userId.ToString());
+
+                await _exceptionLogic.LogException(ex, "Get user review years error", arguments);
+
                 return new List<int>();
             }
         }
@@ -124,6 +164,11 @@ namespace BookReviews.Impl.Logic
         {
             try
             {
+                if (review == null)
+                    throw new Exception("Review is null");
+                else if (!review.IsValid(out string errorMessage))
+                    throw new Exception(errorMessage);
+
                 // first insert the book info if it doesn't exist in the database
                 bool bookInserted = await _bookLogic.AddBook(review.Book);
 
@@ -140,7 +185,15 @@ namespace BookReviews.Impl.Logic
             }
             catch (Exception ex)
             {
-                // log error here
+                var arguments = new Dictionary<string, string>();
+                if (review != null)
+                {
+                    arguments.Add("BookId", review.Book?.Id.ToString());
+                    arguments.Add("UserId", review.User?.Id.ToString());
+                }
+
+                await _exceptionLogic.LogException(ex, "Add review error", arguments);
+
                 return false;
             }
         }
@@ -151,6 +204,11 @@ namespace BookReviews.Impl.Logic
             {
                 // this method assumes that the associated book was successfully inserted
                 // when the review was first written
+                if (review == null)
+                    throw new Exception("Review is null");
+                else if (!review.IsValid(out string errorMessage))
+                    throw new Exception(errorMessage);
+
                 Entities.Review reviewRecord = _mapper.Map<Entities.Review>(review);
                 int recordsUpdated = await _reviewRepository.UpdateReview(reviewRecord);
 
@@ -161,7 +219,16 @@ namespace BookReviews.Impl.Logic
             }
             catch (Exception ex)
             {
-                // log error here
+                var arguments = new Dictionary<string, string>();
+                if (review != null)
+                {
+                    arguments.Add("ReviewId", review.Id.ToString());
+                    arguments.Add("BookId", review.Book?.Id.ToString());
+                    arguments.Add("UserId", review.User?.Id.ToString());
+                }
+
+                await _exceptionLogic.LogException(ex, "Edit review error", arguments);
+
                 return false;
             }
         }
@@ -170,6 +237,9 @@ namespace BookReviews.Impl.Logic
         {
             try
             {
+                if (reviewId == 0)
+                    throw new Exception("Review ID is invalid");
+
                 int recordsDeleted = await _reviewRepository.DeleteReview(reviewId);
 
                 if (recordsDeleted == 0)
@@ -179,7 +249,11 @@ namespace BookReviews.Impl.Logic
             }
             catch (Exception ex)
             {
-                // log error here
+                var arguments = new Dictionary<string, string>();
+                arguments.Add("ReviewId", reviewId.ToString());
+
+                await _exceptionLogic.LogException(ex, "Delete review error", arguments);
+
                 return false;
             }
         }
@@ -195,7 +269,8 @@ namespace BookReviews.Impl.Logic
             }
             catch (Exception ex)
             {
-                // log error here
+                await _exceptionLogic.LogException(ex, "Get genres error");
+
                 return new List<Models.Genre>();
             }
         }
